avcodec/movtextdec: Switch to smaller type
The base size of a box refers to the size the box has in a file,
not in memory; so size_t is not their natural type. Therefore use
a plain unsigned which is smaller on 64bit systems and still big
enough to represent any conceivable base size.

+++ libavcodec/movtextdec.c
@@ -111,7 +111,7 @@ typedef struct {
 
 typedef struct {
     uint32_t type;
-    size_t base_size;
+    unsigned base_size;
     int (*decode)(const uint8_t *tsmb, MovTextContext *m, const AVPacket *avpkt);
 } Box;
